fix(ci): split text diff core module
